Little updates

+++ CarDealer/Models/Contract.cs
@@ -3,7 +3,7 @@ namespace CarDealer.Models;
 public record Contract
 {
 	public int Id { get; set; }
-	public required int Number { get; set; }
+	public required long Number { get; set; }
 	public required DateTime Date { get; set; }
 	public required int Amount { get; set; }
 	public required int ClientId { get; set; }
